% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{Qindex.data-package}
\alias{Qindex.data}
\alias{Qindex.data-package}
\title{Data for Package 'Qindex'}
\description{
Example data used in package 'Qindex'.
}
\references{
Selection of optimal quantile protein biomarkers based on cell-level immunohistochemistry data.
Misung Yi, Tingting Zhan, Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva. BMC Bioinformatics, 2023. \doi{10.1186/s12859-023-05408-8}

Quantile index biomarkers based on single-cell expression data.
Misung Yi, Tingting Zhan, Amy P. Peck, Jeffrey A. Hooke, Albert J. Kovatich, Craig D. Shriver,
Hai Hu, Yunguang Sun, Hallgeir Rui and Inna Chervoneva.
Laboratory Investigation, 2023. \doi{10.1016/j.labinv.2023.100158}
}
\author{
Tingting Zhan, Misung Yi and Inna Chervoneva
}
